Refactoring

+++ test_cases/testBonus.cpp
@@ -1,14 +1,14 @@
 #include <gtest/gtest.h>
 #include <vector>
-#include "../header_files/spieler.h"
+#include "spieler.h"
 
 using namespace std;
 
 TEST(BonusTest, testBonus) {
     Spieler s("Testspieler");
-    for (int i = 0; i < 6; i++) {
-        int z = i + 1;
-        vector<int> v {z, z, z, z, z};
+    for (size_t i = 0; i < 6; i++) {
+        unsigned int z = i + 1;
+        vector<unsigned int> v {z, z, z, z, z};
         s.getKategorienListe().at(i)->punkteBerechnen(v);
     }
     s.bonusErmitteln();
@@ -17,12 +17,11 @@ TEST(BonusTest, testBonus) {
 
 TEST(BonusTest, testKeinBonus) {
     Spieler s("Testspieler");
-    for (int i = 0; i < 6; i++) {
-        int z = i + 1;
-        vector<int> v {z, z, 0, 0, 0};
+    for (size_t i = 0; i < 6; i++) {
+        unsigned int z = i + 1;
+        vector<unsigned int> v {z, z, 0, 0, 0};
         s.getKategorienListe().at(i)->punkteBerechnen(v);
     }
     s.bonusErmitteln();
     EXPECT_EQ(0, s.getPunktzahl());
 }
-
